Modular getters and parsers
