+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "test": "mocha"
+    "test": "mocha --recursive"
   },
   "author": "Viktor Nawrath",
   "license": "ISC",
