+++ src/getters/web.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const utils = require('./utils');
+
+function getData(restaurant) {
+    return utils.getData(restaurant.url);
+};
+
+module.exports = getData;
